:recycle: extracted functions into a function file to use it in several components
